Make ART internals free of 'struct sockaddr'.
Keep route entry/BSD compatibility goos in the rtable layer.  The way
addresses and masks (prefix-lengths) are encoded is really tied to the
radix-tree implementation.

Since we decided to no longer support non-contiguous masks, we could get
rid of some extra "sockaddr" allocations and reduce the memory grows
related to the use of a multibit-trie.