Split gre(4) into two interfaces: gre(4) and mobileip(4).
Like vlan/svlan and tun/tap, it remains a single driver that decides
on the mode based on the interface name.  This removes the need for
removing the default link0 flag to turn gre into Mobile IP mode.
Using linkX / IFF_LINK{0,1,2} for interface modes is a deprecated style.

OK millert@ dlg@ many

Not much consensus on the name but mobileip is also used for the sysctls, so
OK henning@