Add support for packets spanning multiple pages.
The size of TX fragment DMA maps is increased to fit a contiguous
buffer of up to MTU size bytes spanning several pages, e.g. 2 for
4000, 4 for 9000 and so on.

This change also changes the function that calculates an mbuf
chain length so that it tells the driver exactly how many
descriptors will be required in order to provide references to
all buffers within the mbuf chain.