Correct the flag checks inside rt_ifa_addlocal(9) and rt_ifa_dellocal(9).
There's no need to insert an RTF_LOCAL route if it is already there, not
if a route with the same destination exist.

This fixes a KASSERT() triggered by adding an alias for an address already
present in the ARP cache as reported by weerd@ and Peter J. Philipp.

This should also fix a KASSERT() triggered by a NDP change reported by
Sebastien Marie.

ok bluhm@