Use rtalloc(9) instead of ifa_ifwithnet().
ifa_ifwithnet() checks if a given address is directly connected.  This
function predates the introduction of the BSD routing table.  Nowdays
we can check if the route for the given address is marked as RTF_GATEWAY.

This works on OpenBSD because we always install RTF_CONNECTED routes
for subnets a and RTF_HOST route per p2p link.

ok vgross@