New XenStore public API function to read numeric values
A need for a function to perform string to number conversion arose
when domain identifier needed to be read and converted to numerical
representation.  With xbf(4) the usage became broader as greater
values (such as the sector count) needed to be converted.  And as a
result another function was implemented to perform string to unsigned
long long conversion but unfortunately multiplication overflows were
not handled correctly.  This new version consolidates the code in
one place and exports a proper XenStore API function to get and set
numeric values.  The new atoull function borrows multiplication
overflow detection logic from the libc.