Move virtio.c to sys/dev/pv/virtio.c
virtio itself is not PCI, it is a paravirtual bus on top of either PCI
or MMIO (arm).  This is the second step of moving the virtio files to
dev/pv.  No functional change.

Moving the file is done by removing and adding it again - the old
history can be found in the Attic of sys/dev/pci/virtio.c.

no objections from sf@ mlarkin@