On Apple hardware, claim an OSI of Darwin and no other OSes
The DSDT on at MacBooks takes many different and beneficial code
paths when the OS reports to be Darwin, but due to the way it checks
OSI, we can't just add Darwin to the aml_valid_osi array since later
checks for Windows variants will supersede the Darwin check.

Fixes suspend and keeps Thunderbolt ports powered up after resume.
One regression is that the acpibat driver no longer attaches and
will require an ACPI Smart Battery driver, which is in the works.

Adapted from a similar change in the Linux kernel from 2014.

ok deraadt