Put back copying of the response message into the polling callback
The response message is allocated on the stack of hvs_intr and its
content is valid during a single message processing loop iteration.
Due to the fact that polled transfers finish completion in a callee
context, the cached pointer to the response message becomes stale
very quickly.  To avoid this situation the poll completion callback
needs to make sure the content of the response is provided to the
next completion routine.