Use the rv variable in sppp_output() as it was intended.
sppp_output has the following comment:
"Don't choke with an ENETDOWN early.  It's possible that we just
started dialing out, so don't drop the packet immediately.  If we
notice that we run out of buffer space below, we will however remember
that we are not ready to carry IPv6 packets, and return ENETDOWN, as
opposed to ENOBUFS."

A later revision of the code just returned ENOBUFS anyway, ignoring
the rv variable that might have ENETDOWN instead, turning rv into an
unused variable.

Coverity CID 1299246; Severity: Minor

OK sthen@